Extract validation transformer in patient search to separate class
